Echange

+++ EulerProject/src/andy/EulerProject/Problem002.java
@@ -13,8 +13,8 @@ public class Problem002 {
 
 	public static void main(String[] args) {
 		int i = 2;
-		long next = (long) UsefulMethodMath.getFibonacci(i);
-		long sum = 0l;
+		int next = UsefulMethodMath.getFibonacci(i);
+		int sum = 0;
 		
 		while(next <= /*What*/) {
 			//Put Your code here
